RocksDB
Data Verification Addition

+++ src/org/opendedup/collections/RocksDBMap.java
@@ -71,7 +71,7 @@ public class RocksDBMap implements AbstractMap, AbstractHashesMap {
 	int multiplier = 0;
 	boolean closed = false;
 	private long size = 0;
-	private static final long rmthreashold = 5 * 60 * 1000;
+	private static final long rmthreashold = 15 * 60 * 1000;
 	private transient RejectedExecutionHandler executionHandler = new BlockPolicy();
 	private transient BlockingQueue<Runnable> worksQueue = new ArrayBlockingQueue<Runnable>(2);
 	private transient ThreadPoolExecutor executor = null;
@@ -169,7 +169,7 @@ public class RocksDBMap implements AbstractMap, AbstractHashesMap {
 				options.setAllowConcurrentMemtableWrite(true);
 				// LRUCache c = new LRUCache(memperDB);
 				// options.setRowCache(c);
-				//blockConfig.setBlockCacheSize(GB * 2);
+				// blockConfig.setBlockCacheSize(GB * 2);
 
 				options.setWriteBufferSize(bufferSize / dbs.length);
 				// options.setMinWriteBufferNumberToMerge(2);
@@ -298,12 +298,13 @@ public class RocksDBMap implements AbstractMap, AbstractHashesMap {
 							+ "] hash locations didn't match stored val=" + oval + " request value=" + val);
 					return false;
 				}
+				long oct = ct;
 				ct += bk.getLong();
-				if (ct <= 0) {
+				if (ct <= 0 && oct < 0) {
 					bk.position(8);
 					bk.putLong(System.currentTimeMillis() + rmthreashold);
 					rmdb.put(hash, v);
-				} else if (rmdb.get(hash) != null) {
+				} else if(rmdb.get(hash) != null) {
 					rmdb.delete(hash);
 				}
 				bk.putLong(v.length - 8, ct);
@@ -374,33 +375,36 @@ public class RocksDBMap implements AbstractMap, AbstractHashesMap {
 				byte[] hash = iter.key();
 				Lock l = this.getLock(hash);
 				l.lock();
+
 				try {
-					byte[] v = null;
-					bk.position(0);
-					bk.put(iter.value());
-					bk.position(0);
-					long pos = bk.getLong();
-					long tm = bk.getLong();
-					
-					if (System.currentTimeMillis() > tm) {
-						v = this.getDB(hash).get(hash);
-						if (v != null) {
-							ByteBuffer nbk = ByteBuffer.wrap(v);
-							long oval = nbk.getLong();
-							long ct = nbk.getLong();
-							if (ct <= 0 && oval == pos) {
+					if (this.rmdb.get(hash) != null) {
+						byte[] v = null;
+						bk.position(0);
+						bk.put(iter.value());
+						bk.position(0);
+						long pos = bk.getLong();
+						long tm = bk.getLong();
+
+						if (System.currentTimeMillis() > tm) {
+							v = this.getDB(hash).get(hash);
+							if (v != null) {
+								ByteBuffer nbk = ByteBuffer.wrap(v);
+								long oval = nbk.getLong();
+								long ct = nbk.getLong();
+								if (ct <= 0 && oval == pos) {
+									ChunkData ck = new ChunkData(pos, iter.key());
+									ck.setmDelete(true);
+									this.getDB(hash).delete(hash);
+								}
+								rmdb.delete(iter.key());
+							} else {
+								rmdb.delete(iter.key());
 								ChunkData ck = new ChunkData(pos, iter.key());
 								ck.setmDelete(true);
-								this.getDB(hash).delete(hash);
-							}
-							rmdb.delete(iter.key());
-						} else {
-							rmdb.delete(iter.key());
-							ChunkData ck = new ChunkData(pos, iter.key());
-							ck.setmDelete(true);
 
+							}
+							rmk++;
 						}
-						rmk++;
 					}
 				} finally {
 					l.unlock();
@@ -517,12 +521,16 @@ public class RocksDBMap implements AbstractMap, AbstractHashesMap {
 					else
 						bf.putLong(cm.references);
 					db.put(cm.getHash(), v);
+					this.rmdb.delete(cm.getHash());
 					return new InsertRecord(true, cm.getcPos());
 				} else {
 					// SDFSLogger.getLog().info("Hash Found");
 					ByteBuffer bk = ByteBuffer.wrap(v);
 					long pos = bk.getLong();
 					long ct = bk.getLong();
+					if (ct <= 0) {
+						ct = 0;
+					}
 					if (cm.references <= 0)
 						ct++;
 					else
@@ -639,8 +647,21 @@ public class RocksDBMap implements AbstractMap, AbstractHashesMap {
 				SDFSLogger.getLog().warn("unable to get key [" + StringUtils.getHexString(key) + "] [" + pos + "]", e);
 			}
 			if (direct && (data == null || data.length == 0)) {
+				if (data != null && data.length == 0) {
+					SDFSLogger.getLog().warn("Data Lenght is 0");
+				}
 				SDFSLogger.getLog().warn(" miss for [" + StringUtils.getHexString(key) + "] [" + pos + "] ");
-				return null;
+				long opos = pos;
+				pos = this.get(key);
+				if (pos != -1) {
+					SDFSLogger.getLog().warn(
+							" miss for [" + StringUtils.getHexString(key) + "] [" + opos + "] found at [" + pos + "]");
+					return data = ChunkData.getChunk(key, pos);
+				} else {
+					SDFSLogger.getLog()
+							.warn(" do data found for [" + StringUtils.getHexString(key) + "] [" + pos + "]");
+					return null;
+				}
 			} else {
 				return data;
 			}
@@ -775,7 +796,7 @@ public class RocksDBMap implements AbstractMap, AbstractHashesMap {
 	}
 
 	@Override
-	public boolean mightContainKey(byte[] key,long id) {
+	public boolean mightContainKey(byte[] key, long id) {
 		Lock l = this.getLock(key);
 		l.lock();
 		try {
@@ -794,7 +815,6 @@ public class RocksDBMap implements AbstractMap, AbstractHashesMap {
 		}
 	}
 
-
 	public static class StartShard implements Runnable {
 
 		RocksDB[] dbs = null;
@@ -805,9 +825,7 @@ public class RocksDBMap implements AbstractMap, AbstractHashesMap {
 		CommandLineProgressBar bar = null;
 		AtomicInteger ct = null;
 
-
-		public StartShard(int n, RocksDB[] dbs, Options options, File f, CommandLineProgressBar bar,
-				AtomicInteger ct) {
+		public StartShard(int n, RocksDB[] dbs, Options options, File f, CommandLineProgressBar bar, AtomicInteger ct) {
 			this.dbs = dbs;
 			this.n = n;
 			this.options = options;
